feat: remove WS path
