+++ package.json
@@ -47,7 +47,6 @@
     "signale": "^1.4.0",
     "transliteration": "^2.3.5",
     "unzipper": "^0.12.3",
-    "ws": "^8.20.0",
     "yauzl-promise": "^4.0.0",
     "zod": "^3.22.4"
   },
@@ -62,7 +61,6 @@
     "@types/nconf": "^0.10.6",
     "@types/signale": "^1.4.7",
     "@types/unzipper": "^0.10.11",
-    "@types/ws": "^8.18.1",
     "@types/yauzl-promise": "^4.0.1",
     "concurrently": "^8.2.2",
     "esbuild": "^0.25.2",
